requirements.txt
pandas
boto3
psycopg2-binary

google-api-python-client
gspread
google-auth

pyarrow
fastparquet

python-dotenv
requests